chore: add useSubplotExtent custom hook
