+++ v3/src/components/graph/adornments/adornment.tsx
@@ -3,8 +3,7 @@ import { clsx } from "clsx"
 import { observer } from "mobx-react-lite"
 import { IAdornmentModel } from "./adornment-models"
 import { useGraphContentModelContext } from "../hooks/use-graph-content-model-context"
-import { useGraphLayoutContext } from "../hooks/use-graph-layout-context"
-import { useGraphDataConfigurationContext } from "../hooks/use-graph-data-configuration-context"
+import { useSubplotExtent } from "../hooks/use-subplot-extent"
 import { INumericAxisModel } from "../../axis/models/axis-model"
 import { getAdornmentComponentInfo } from "./adornment-component-info"
 import {transitionDuration} from "../../data-display/data-display-types"
@@ -18,12 +17,7 @@ interface IProps {
 
 export const Adornment = observer(function Adornment({adornment, cellKey}: IProps) {
   const graphModel = useGraphContentModelContext()
-  const layout = useGraphLayoutContext()
-  const dataConfig = useGraphDataConfigurationContext()
-  const numExtraPrimaryBands = dataConfig?.numRepetitionsForPlace("bottom") ?? 1
-  const numExtraSecondaryBands = dataConfig?.numRepetitionsForPlace("left") ?? 1
-  const subPlotWidth = layout.plotWidth / numExtraPrimaryBands
-  const subPlotHeight = layout.plotHeight / numExtraSecondaryBands
+  const { subPlotWidth, subPlotHeight } = useSubplotExtent()
   const classFromCellKey = adornment.classNameFromKey(cellKey)
   // The adornmentKey is a unique value used for React's key prop and for the adornment wrapper's HTML ID.
   // We can't use the cellKey because that value may be duplicated if there are multiple types of
